+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.1.0
+  Add method to get all keys from the cache.
+  Add method to delete a specific or a pattern of keys from the cache.
+
 ## 0.0.1
   Initial release.